Remove requirement for segmenting syllables
To discretize spectrograms, it was required to use an amplitude threshold to segment syllables. Now the threshold can be set to 0 if the user does not want to segment syllables. Before, the user could set a threshold of 0, but setting the parameter to this value could lead to aberrant behavior when the recording is noisy.